requirements.txt
FLASK==1.1.1
Werkzeug==0.16.0
gunicorn==20.0.0
numpy>=1.9.2
Jinja2==2.10.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
scipy>=0.15.1
scikit_learn>=0.18
pandas>=0.19
nltk>=3.4.5
newspaper3k

Flask-Cors==3.0.8
requests
gevent